requirements.txt
h5py==3.7.0
ipython==8.7.0
keras==2.11.0
matplotlib==3.6.3
numpy==1.17.4
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.68
Pillow==9.5.0
scikit_learn==1.2.0
tensorflow==2.11.0
tensorflow_gpu==2.11.0
